Change test strings for clearer picture of what's going on
It was unclear to me what scenarios would legitimately render a body that matched "error" vs when "hello world" was rendered because index.html was rendered or because the middleware was rendering that string directly.

This changes index.html to contain the string "Original body" and the middleware to send the string "special error handling" to disambiguate the two.

Additionally it adds some assertions around what the body _should_ match for examples that only had what it _should not_ match.